qwen3.5b-on-lambda/src/client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.59",
    "botocore[crt]>=1.42.59",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
